Enable and configure OpsGenie alerts for production environment
This update adds support for OpsGenie alerts, specifically for the production environment. It introduces a new variable for the monitor resource group name, a local value for the OpsGenie action group, and modifies the alert configurations to include OpsGenie if the environment is production. Additionally, alerts have been enabled in the "prod01" environment configuration.